requirements.txt
streamlit-nightly
openai

PyPDF2